requirements.txt
streamlit
groq
PyPDF2
sentence-transformers
faiss-cpu
python-dotenv
numpy